+++ src/test/groovy/ru/vyarus/dropwizard/guice/OptionsLookupTest.groovy
@@ -23,7 +23,7 @@ class OptionsLookupTest extends AbstractTest {
     def "Test options multi config"() {
 
         expect: "options set"
-        info.options.getValue(GuiceyOptions.BindConfigurationInterfaces)
+        !info.options.getValue(GuiceyOptions.UseCoreInstallers)
     }
 
     static class OLApp extends Application<Configuration> {
@@ -31,7 +31,7 @@ class OptionsLookupTest extends AbstractTest {
         void initialize(Bootstrap<Configuration> bootstrap) {
             bootstrap.addBundle(GuiceBundle.builder()
                     .options(ImmutableMap.<Enum, Object> builder()
-                        .put(GuiceyOptions.BindConfigurationInterfaces, true)
+                        .put(GuiceyOptions.UseCoreInstallers, false)
                         .build())
                     .build())
         }