+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/debug/renderer/OptionsRendererTest.groovy
@@ -45,7 +45,6 @@ class OptionsRendererTest extends Specification {
         ScanPackages                   = [com.foo, com.bat]             *CUSTOM
         SearchCommands                 = false
         UseCoreInstallers              = true
-        BindConfigurationInterfaces    = false
         BindConfigurationByPath        = true
         InjectorStage                  = PRODUCTION
         GuiceFilterRegistration        = [REQUEST]
@@ -79,7 +78,6 @@ class OptionsRendererTest extends Specification {
         ScanPackages                   = [com.foo, com.bat]             *CUSTOM
         SearchCommands                 = false
         UseCoreInstallers              = true
-        BindConfigurationInterfaces    = false
         BindConfigurationByPath        = true
         InjectorStage                  = PRODUCTION
         GuiceFilterRegistration        = [REQUEST]
@@ -132,7 +130,6 @@ class OptionsRendererTest extends Specification {
         ScanPackages                   = [com.foo, com.bat]             *CUSTOM
         SearchCommands                 = false
         UseCoreInstallers              = true
-        BindConfigurationInterfaces    = false
         BindConfigurationByPath        = true
         InjectorStage                  = PRODUCTION
         UseHkBridge                    = false