+++ CHANGELOG.md
@@ -23,7 +23,8 @@
         - Extensions classpath search and validation (but on run phase it is still possible to disable extensions)    
     - A lot of guicey lifecycle events obviously changed        
     - Removed GuiceyOptions.ConfigureFromDropwizardBundles option because it's useless with new bundles lifecycle.
-        (if required, the same behaviour may be implemented with custom bundles lookup)                  
+        (if required, the same behaviour may be implemented with custom bundles lookup)
+* Remove GuiceyOptions.BindConfigurationInterfaces option (interfaces are already bound with @Config qualifier)                          
 
 Main breaking changes were caused by:
  - jersey 2.26 introduces an abstraction for injection layer in order to get rid of hk2 direct usage.